# classlite-api/api.yaml
openapi: "3.1.0"
info:
  title: ClassLite API
  version: "0.1.0"
  description: Multi-tenant education management platform API
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/knowledge-hub/folders:
    get:
      operationId: listFolders
      summary: List all (non-deleted) folders for the caller's center — the 4.4b tree source (story 4.4a — AC2).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Flat list of folders (client composes the tree via parentFolderId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFolderList'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      operationId: createFolder
      summary: Create a folder (optionally nested via parentFolderId) (story 4.4a — AC2).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFolder'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FOLDER_NOT_FOUND (parentFolderId absent or in another tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "422":
          description: FOLDER_MAX_DEPTH (nesting beyond the depth limit) / validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/knowledge-hub/folders/{id}:
    patch:
      operationId: updateFolder
      summary: Rename and/or move (reparent) a folder. A move into the folder's own descendant → 422 FOLDER_CYCLE (story 4.4a — AC2).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolderRequest'
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFolder'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FOLDER_NOT_FOUND (absent, soft-deleted, or cross-tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "422":
          description: FOLDER_CYCLE (move into own descendant) / FOLDER_MAX_DEPTH / validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteFolder
      summary: Soft-delete a folder (deleted_at stamp) (story 4.4a — AC2).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Soft-deleted (no body)
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FOLDER_NOT_FOUND (absent or already deleted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/knowledge-hub/files:
    get:
      operationId: listFiles
      summary: List (non-deleted) files in a folder (or root when folderId is omitted) (story 4.4a — AC3).
      security:
        - bearerAuth: []
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Folder to list. Omitted → root-level files (folder_id IS NULL).
      responses:
        "200":
          description: Files in the folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFileList'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/knowledge-hub/files/{slug}:
    get:
      operationId: getFileDetail
      summary: File detail — type-tagged metadata + linked locations (sessions/exercises referencing it). No view-rate (deferred) (story 4.4a — AC13).
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File detail with linked locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFileDetail'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FILE_NOT_FOUND (absent, soft-deleted, or cross-tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/knowledge-hub/files/{slug}/download:
    get:
      operationId: getFileDownloadUrl
      summary: >-
        Short-lived presigned GET URL for a file's stored object (story 4.4b —
        AC5 preview/download). The file is resolved by slug within the caller's
        tenant (RLS); the object key's {center_id} prefix is re-asserted (SEC-8)
        before the GET is signed. Expiry 5 min. URL never logged (A10). With
        `disposition=attachment` the signed URL forces a download with the
        file's original name (Content-Disposition: attachment); omitted, the URL
        serves inline so the same URL can back the `<img>`/`<audio>`/`<embed>`
        preview.
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: disposition
          in: query
          required: false
          description: >-
            `attachment` forces a download (Content-Disposition: attachment;
            filename=<original name>). Omitted → inline, for preview.
          schema:
            type: string
            enum: [attachment]
      responses:
        "200":
          description: Presigned GET URL for the object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeDownloadUrl'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / R2_KEY_PREFIX_MISMATCH / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FILE_NOT_FOUND (absent, soft-deleted, or cross-tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/knowledge-hub/files/{id}:
    patch:
      operationId: updateFile
      summary: Rename and/or move (reparent) a file (story 4.4a — AC3).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeFile'
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FILE_NOT_FOUND / FOLDER_NOT_FOUND (target parent absent or cross-tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "422":
          description: Validation error (name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteFile
      summary: Soft-delete a file (deleted_at stamp; R2 object + row retained; frees storage accounting) (story 4.4a — AC3).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Soft-deleted (no body)
        "401":
          description: AUTH_REQUIRED / AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: FORBIDDEN / EMAIL_VERIFICATION_REQUIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "404":
          description: FILE_NOT_FOUND (absent or already deleted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        code:
          type: string
          description: |
            Optional per-field UPPER_SNAKE_CASE identifier (e.g.
            INVALID_TEACHER_EMAIL, SELF_INVITE_BLOCKED) — populated for
            error-catalog fields; empty string OR omitted when the error
            carries only a prose message. Story 1.x handlers pre-date the
            catalog and emit `""` for `code`; Story 2.2+ handlers populate
            it where the wizard router needs to key on the field-level
            code. C2-03 review fix — was `required` under Story 2.2's
            initial pass, which broke JSON-Schema validators for pre-2.2
            envelopes that didn't emit the key.
        message:
          type: string
    ErrorBody:
      type: object
      required: [code, message, requestId, details]
      properties:
        code:
          type: string
          description: UPPER_SNAKE_CASE error code (e.g., VALIDATION_ERROR, EMAIL_ALREADY_REGISTERED, VERIFICATION_TOKEN_EXPIRED).
        message:
          type: string
        requestId:
          type: string
        details:
          oneOf:
            - type: 'null'
            - type: array
              items:
                $ref: '#/components/schemas/FieldError'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'

    # ---------------- Story 1.6 — Google OAuth + Invite Acceptance + Force-Logout ----------------
...
    File:
      type: object
      # All fields required + explicit (nulls not omitted) per GO-5.
      required:
        - id
        - centerId
        - folderId
        - name
        - slug
        - objectKey
        - contentType
        - sizeBytes
        - uploadedBy
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        centerId:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        slug:
          type: string
          description: Unique-per-center URL-safe identifier for the file detail route.
        objectKey:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        uploadedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LinkedLocation:
      type: object
      required: [type, id, label]
      properties:
        type:
          type: string
          enum: [session, exercise]
        id:
          type: string
          format: uuid
        label:
          type: string
          description: Human-readable label (session title / exercise title) for the linking host.
    FileDetail:
      type: object
      # File + linkedLocations. NO view-rate field (deferred to the analytics epic).
      required:
        - id
        - centerId
        - folderId
        - name
        - slug
        - objectKey
        - contentType
        - sizeBytes
        - uploadedBy
        - createdAt
        - updatedAt
        - linkedLocations
      properties:
        id:
          type: string
          format: uuid
        centerId:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        slug:
          type: string
        objectKey:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        uploadedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        linkedLocations:
          type: array
          items:
            $ref: '#/components/schemas/LinkedLocation'
          description: Sessions/exercises referencing this file. Excludes any whose host or the file itself is soft-deleted.
    Folder:
      type: object
      required:
        - id
        - centerId
        - parentFolderId
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        centerId:
          type: string
          format: uuid
        parentFolderId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parentFolderId:
          type: string
          format: uuid
          nullable: true
          description: Parent folder. null/absent = root-level.
    UpdateFolderRequest:
      type: object
      description: |
        Rename and/or move a folder. Tri-state on parentFolderId:
        key absent = unchanged; explicit null = move to root; a uuid = reparent.
        A move into the folder's own descendant is rejected 422 FOLDER_CYCLE.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parentFolderId:
          type: string
          format: uuid
          nullable: true
    UpdateFileRequest:
      type: object
      description: |
        Rename and/or move a file. Tri-state on folderId: key absent = unchanged;
        explicit null = move to root; a uuid = reparent into that folder.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        folderId:
          type: string
          format: uuid
          nullable: true
...
    EnvelopeFile:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/File'
    EnvelopeFileList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/File'
    EnvelopeFileDetail:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/FileDetail'
    EnvelopeFolder:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Folder'
    EnvelopeFolderList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
...
    DownloadUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Short-lived (5-min) presigned GET URL for the file's stored object. Never logged (A10 slog redaction).
    EnvelopeDownloadUrl:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/DownloadUrl'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT